Tests clean-up
